added sfx

+++ src/com/company/engine/Buffer.java
@@ -45,8 +45,8 @@ public class Buffer {
         graphics.fillRect((int) coord.getX(), (int) coord.getY(), (int) width, (int) height);
     }
 
-    public void drawImage(Image image, int x, int y) {
-        graphics.drawImage(image, x, y, null);
+    public void drawImage(Image image, float x, float y) {
+        graphics.drawImage(image, (int) x, (int) y, null);
     }
 
     public void drawImage(Image image, Point coord) {
